-Finished 8-2 -Fixed some misc. things  (now everything compiles)

+++ 4-6 (String umgekehrt ausgeben; mit Leerzeichen)/main.cpp
@@ -11,7 +11,7 @@ int main()
 	string output = "";
 	if (input.size() > 0)
 	{
-		for (int i = input.size(); i > 0; i--) {
+		for (int i = (int)input.size(); i > 0; i--) {
 			if (!isalpha(input.at(i - 1))) {
 				output.append(" ");
 				continue;
